Moving gen_items in test_utils, updating mods
