+++ src/lib.rs
@@ -1,5 +1,6 @@
 //! An implementation of Approximate Lower Bound Arguments
 //! (ALBA, <https://eprint.iacr.org/2023/1655.pdf>).
 pub mod utils;
+pub mod test_utils;
 
 pub mod centralized;
