+++ src/test_utils.rs
@@ -1,0 +1,16 @@
+use crate::utils;
+
+// Test & Bench helpers
+
+/// Generate a set of items given the set size and a seed
+/// Items are generated by hashing the current index
+pub fn gen_items<const N: usize>(seed: Vec<u8>, set_size: usize) -> Vec<[u8; N]> {
+    let mut s_p = Vec::with_capacity(set_size);
+    for b in 0..set_size {
+        let mut data = vec![seed.clone()];
+        data.push(b.to_ne_bytes().to_vec());
+        let item = utils::combine_hashes::<N>(data);
+        s_p.push(item);
+    }
+    s_p
+}
